requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sqlalchemy>=2.0
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.50.0
tiktoken>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
structlog>=24.0.0
